VIH-4577: POC WIP
