+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/app.module.ts
@@ -38,7 +38,7 @@ import { BaseJourneyModule } from './modules/base-journey/base-journey.module';
 import { SecurityModule } from './modules/security/security.module';
 import { SharedModule } from './modules/shared/shared.module';
 import { SERVICE_WEB_API_BASE_URL } from './services/clients/api-client';
-import {SelfTestModule} from './modules/self-test/self-test.module';
+import {SelfTestModule} from './modules/self-test-journey/self-test.module';
 
 export function initConfiguration(configService: ConfigService): Function {
   return () => configService.load();
