+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/self-test-journey/pages/self-test-base-component/self-test-base.component.ts
