+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/self-test-journey/paths.ts
