+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/self-test-journey/self-test-journey-steps.ts
@@ -3,4 +3,5 @@ import { ParticipantJourneySteps } from '../base-journey/participant-journey-ste
 
 export class SelfTestJourneySteps extends ParticipantJourneySteps {
   static readonly  First = new JourneyStep('First');
+  static readonly  Dropout = new JourneyStep('Dropout');
 }
