+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/self-test-journey/self-test-steps-order.factory.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { DeviceType } from '../../modules/base-journey/services/device-type';
-import { IndividualJourneySteps } from '../individual-journey/individual-journey-steps';
+import { SelfTestJourneySteps } from './self-test-journey-steps';
 import { JourneyStep } from '../base-journey/journey-step';
 
 
 @Injectable()
-export class IndividualStepsOrderFactory {
+export class SelfTestStepsOrderFactory {
 
   constructor(private deviceType: DeviceType) {
   }
@@ -13,11 +13,11 @@ export class IndividualStepsOrderFactory {
   stepOrder(): JourneyStep[] {
     if (this.deviceType.isMobile()) {
       return [
-        IndividualJourneySteps.AboutHearings
+        SelfTestJourneySteps.First
       ];
     } else {
       return [
-        IndividualJourneySteps.AboutHearings
+        SelfTestJourneySteps.First
       ];
     }
   }
