+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/self-test-journey/self-test.model.ts
